requirements.txt
autopep8==1.6.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
config==0.5.1
cryptography==36.0.2
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
virtualenv==20.14.1
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
